+++ NoneBot/src/plugins/siyuan/plugins/account/bind.py
@@ -1,27 +1,29 @@
 # Copyright (C) 2023 Zuoqiu Yingyi
-#
+# 
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as
 # published by the Free Software Foundation, either version 3 of the
 # License, or (at your option) any later version.
-#
+# 
 # This program is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU Affero General Public License for more details.
-#
+# 
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from nonebot.plugin import PluginMetadata
 
-__plugin_meta__ = PluginMetadata(
-    name="account",
-    description="账户管理",
-    usage="",
+from nonebot import on_command
+from nonebot.adapters.onebot.v11 import (
+    Bot,
+    MessageEvent,
 )
 
-
-from . import key
-from . import user
-from . import bind
+accound_bind = on_command(
+    cmd="bind",
+    aliases={
+        "绑定",
+    },
+    block=True,
+)