+++ NoneBot/src/plugins/siyuan/plugins/account/key.py
@@ -13,15 +13,26 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from nonebot.plugin import PluginMetadata
+from nonebot import on_command
+from nonebot.adapters.onebot.v11 import (
+    Bot,
+    MessageEvent,
+)
+from ....siyuan import pgp
 
-__plugin_meta__ = PluginMetadata(
-    name="account",
-    description="账户管理",
-    usage="",
+public_key = on_command(
+    cmd="key",
+    aliases={
+        "公钥",
+        "PGP公钥",
+    },
+    block=True,
 )
 
 
-from . import key
-from . import user
-from . import bind
+@public_key.handle()
+async def _(
+    bot: Bot,
+    event: MessageEvent,
+):
+    await public_key.finish(f"PGP 公钥：\n\n{pgp.encrypt_key.pubkey}")